cleanup fortran compiler code use size_t where possible instead of uint type

+++ bld/f77/wfc/c/iflookup.c
@@ -2,7 +2,7 @@
 *
 *                            Open Watcom Project
 *
-* Copyright (c) 2002-2020 The Open Watcom Contributors. All Rights Reserved.
+* Copyright (c) 2002-2021 The Open Watcom Contributors. All Rights Reserved.
 *    Portions Copyright (c) 1983-2002 Sybase, Inc. All Rights Reserved.
 *
 *  ========================================================================
@@ -91,8 +91,8 @@ IFF     IFLookUp( void ) {
 }
 
 
-static sym_id IFSymLookup( const char *name, uint len )
-//=====================================================
+static sym_id IFSymLookup( const char *name, size_t len )
+//=======================================================
 {
     sym_id      sym;
